+++ editor/app/(dev)/canvas/examples/with-templates/002/page.tsx
@@ -3,7 +3,7 @@
 import React from "react";
 import { editor } from "@/grida-canvas";
 import Editor from "../../../editor";
-import queryattributes from "@/grida-canvas-react/nodes/utils/attributes";
+import queryattributes from "@/grida-canvas-react-renderer-dom/nodes/utils/attributes";
 import ReferrerPageTemplate from "@/theme/templates/west-referral/referrer/page";
 
 const document: editor.state.IEditorStateInit = {
