+++ editor/app/(dev)/canvas/experimental/wasm-test/page.tsx
@@ -1,7 +1,7 @@
 "use client";
 import * as React from "react";
 import init, { type Grida2D } from "@grida/canvas-wasm";
-import locateFile from "@/grida-canvas-wasm-react/locate-file";
+import locateFile from "@/grida-canvas-react-renderer-canvas-wasm/locate-file";
 
 const __test_document = `{"version":"0.0.1-beta.1+20250303","document":{"bitmaps":{},"images":{},"properties":{},"nodes":{"25575e75-a544-4fa3-b199-15d1906588b2":{"id":"25575e75-a544-4fa3-b199-15d1906588b2","name":"rectangle","locked":false,"active":true,"position":"absolute","top":0,"left":0,"opacity":1,"zIndex":0,"rotation":0,"fill":{"type":"solid","color":{"r":217,"g":217,"b":217,"a":1}},"width":100,"height":100,"style":{},"type":"rectangle","cornerRadius":0,"effects":[],"strokeWidth":0,"strokeCap":"butt"}},"scenes":{"main":{"type":"scene","guides":[],"edges":[],"constraints":{"children":"multiple"},"children":["25575e75-a544-4fa3-b199-15d1906588b2"],"id":"main","name":"main","backgroundColor":{"r":245,"g":245,"b":245,"a":1}}}}}`;
