+++ editor/grida-canvas-react-renderer-dom/nodes/path.tsx
