+++ editor/grida-canvas-react-renderer-dom/nodes/vector.tsx
