+++ editor/theme/templates/formcollection/page.tsx
@@ -3,14 +3,14 @@
 import React from "react";
 import { PoweredByGridaFooter } from "@/grida-forms-hosted/e/powered-by-brand-footer";
 import { ToggleGroup, ToggleGroupItem } from "@/components/ui/toggle-group";
-import { NodeElement } from "@/grida-canvas-react/nodes/node";
+import { NodeElement } from "@/grida-canvas-react-renderer-dom/nodes/node";
 import {
   Card_002,
   Card_003,
   Hero_002,
-} from "@/grida-canvas-react/template-builder/components/cards";
-import { Footer_001 } from "@/grida-canvas-react/template-builder/components/footers";
-import { Header_001 } from "@/grida-canvas-react/template-builder/components/headers";
+} from "@/grida-canvas-react-renderer-dom/template-builder/components/cards";
+import { Footer_001 } from "@/grida-canvas-react-renderer-dom/template-builder/components/footers";
+import { Header_001 } from "@/grida-canvas-react-renderer-dom/template-builder/components/headers";
 import * as samples from "./samples";
 import {
   ProgramDataContextHost,
