+++ editor/theme/templates/formstart/003/page.tsx
@@ -9,7 +9,7 @@ import {
   TextAlign,
 } from "@/theme/templates/kit/components";
 import type grida from "@grida/schema";
-import { NodeElement } from "@/grida-canvas-react/nodes/node";
+import { NodeElement } from "@/grida-canvas-react-renderer-dom/nodes/node";
 import { useCTAContext } from "../../kit/contexts/cta.context";
 import { factory } from "@grida/tokens";
